+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/grid/WirelessGridContainerMenu.java
@@ -1,7 +1,7 @@
 package com.refinedmods.refinedstorage.common.grid;
 
 import com.refinedmods.refinedstorage.common.api.grid.Grid;
-import com.refinedmods.refinedstorage.common.api.support.slotreference.SlotReference;
+import com.refinedmods.refinedstorage.common.api.support.slotreference.PlayerSlotReference;
 import com.refinedmods.refinedstorage.common.content.Menus;
 
 import net.minecraft.world.entity.player.Inventory;
@@ -11,16 +11,16 @@ public class WirelessGridContainerMenu extends AbstractGridContainerMenu {
                                      final Inventory playerInventory,
                                      final WirelessGridData wirelessGridData) {
         super(Menus.INSTANCE.getWirelessGrid(), syncId, playerInventory, wirelessGridData.gridData());
-        this.disabledSlot = wirelessGridData.slotReference();
+        this.disabledSlot = wirelessGridData.playerSlotReference();
         resized(0, 0, 0);
     }
 
     WirelessGridContainerMenu(final int syncId,
                               final Inventory playerInventory,
                               final Grid grid,
-                              final SlotReference slotReference) {
+                              final PlayerSlotReference playerSlotReference) {
         super(Menus.INSTANCE.getWirelessGrid(), syncId, playerInventory, grid);
-        this.disabledSlot = slotReference;
+        this.disabledSlot = playerSlotReference;
         resized(0, 0, 0);
     }
 }