+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/grid/WirelessGridItem.java
@@ -7,11 +7,13 @@ import com.refinedmods.refinedstorage.common.api.RefinedStorageApi;
 import com.refinedmods.refinedstorage.common.api.grid.Grid;
 import com.refinedmods.refinedstorage.common.api.security.SecurityHelper;
 import com.refinedmods.refinedstorage.common.api.support.energy.AbstractNetworkEnergyItem;
+import com.refinedmods.refinedstorage.common.api.support.energy.EnergyItemContext;
 import com.refinedmods.refinedstorage.common.api.support.network.item.NetworkItemContext;
-import com.refinedmods.refinedstorage.common.api.support.slotreference.SlotReference;
+import com.refinedmods.refinedstorage.common.api.support.slotreference.PlayerSlotReference;
 import com.refinedmods.refinedstorage.common.content.ContentIds;
 import com.refinedmods.refinedstorage.common.content.ContentNames;
 import com.refinedmods.refinedstorage.common.security.BuiltinPermission;
+import com.refinedmods.refinedstorage.common.support.energy.ItemEnergyStorage;
 
 import net.minecraft.core.registries.Registries;
 import net.minecraft.network.chat.Component;
@@ -36,17 +38,17 @@ public class WirelessGridItem extends AbstractNetworkEnergyItem {
         this.creative = creative;
     }
 
-    public EnergyStorage createEnergyStorage(final ItemStack stack) {
+    public static EnergyStorage createEnergyStorage(final ItemStack stack, final EnergyItemContext context) {
         final EnergyStorage energyStorage = new EnergyStorageImpl(
             Math.clamp(Platform.INSTANCE.getConfig().getWirelessGrid().getEnergyCapacity(), 1, Long.MAX_VALUE)
         );
-        return RefinedStorageApi.INSTANCE.asItemEnergyStorage(energyStorage, stack);
+        return new ItemEnergyStorage(stack, energyStorage, context);
     }
 
     @Override
     protected void use(@Nullable final Component name,
                        final ServerPlayer player,
-                       final SlotReference slotReference,
+                       final PlayerSlotReference playerSlotReference,
                        final NetworkItemContext context) {
         final boolean isAllowed = context.resolveNetwork()
             .map(network -> SecurityHelper.isAllowed(player, BuiltinPermission.OPEN, network))
@@ -60,7 +62,7 @@ public class WirelessGridItem extends AbstractNetworkEnergyItem {
             name,
             creative ? ContentNames.CREATIVE_WIRELESS_GRID : ContentNames.WIRELESS_GRID
         );
-        final var provider = new WirelessGridExtendedMenuProvider(correctedName, grid, slotReference);
+        final var provider = new WirelessGridExtendedMenuProvider(correctedName, grid, playerSlotReference);
         Platform.INSTANCE.getMenuOpener().openMenu(player, provider);
     }
 }