+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/security/SecurityCardData.java
@@ -2,8 +2,7 @@ package com.refinedmods.refinedstorage.common.security;
 
 import com.refinedmods.refinedstorage.common.api.RefinedStorageApi;
 import com.refinedmods.refinedstorage.common.api.security.PlatformPermission;
-import com.refinedmods.refinedstorage.common.api.support.slotreference.SlotReference;
-import com.refinedmods.refinedstorage.common.api.support.slotreference.SlotReferenceFactory;
+import com.refinedmods.refinedstorage.common.api.support.slotreference.PlayerSlotReference;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -12,9 +11,9 @@ import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraft.network.codec.ByteBufCodecs;
 import net.minecraft.network.codec.StreamCodec;
 
-public record SecurityCardData(SlotReference slotReference, List<Permission> permissions) {
+public record SecurityCardData(PlayerSlotReference playerSlotReference, List<Permission> permissions) {
     public static final StreamCodec<RegistryFriendlyByteBuf, SecurityCardData> STREAM_CODEC = StreamCodec.composite(
-        SlotReferenceFactory.STREAM_CODEC, SecurityCardData::slotReference,
+        PlayerSlotReference.STREAM_CODEC, SecurityCardData::playerSlotReference,
         ByteBufCodecs.collection(
             ArrayList::new,
             StreamCodec.composite(