+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/support/packet/c2s/UseSlotReferencedItemPacket.java
@@ -1,8 +1,7 @@
 package com.refinedmods.refinedstorage.common.support.packet.c2s;
 
-import com.refinedmods.refinedstorage.common.api.support.slotreference.SlotReference;
-import com.refinedmods.refinedstorage.common.api.support.slotreference.SlotReferenceFactory;
-import com.refinedmods.refinedstorage.common.api.support.slotreference.SlotReferenceHandlerItem;
+import com.refinedmods.refinedstorage.common.api.support.slotreference.PlayerSlotReference;
+import com.refinedmods.refinedstorage.common.api.support.slotreference.UsablePlayerSlotReferencedItem;
 import com.refinedmods.refinedstorage.common.support.packet.PacketContext;
 
 import net.minecraft.network.RegistryFriendlyByteBuf;
@@ -10,27 +9,27 @@ import net.minecraft.network.codec.StreamCodec;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.ItemStack;
 
 import static com.refinedmods.refinedstorage.common.util.IdentifierUtil.createIdentifier;
 
-public record UseSlotReferencedItemPacket(SlotReference slotReference) implements CustomPacketPayload {
+public record UseSlotReferencedItemPacket(PlayerSlotReference playerSlotReference) implements CustomPacketPayload {
     public static final Type<UseSlotReferencedItemPacket> PACKET_TYPE = new Type<>(
         createIdentifier("use_slot_referenced_item")
     );
     public static final StreamCodec<RegistryFriendlyByteBuf, UseSlotReferencedItemPacket> STREAM_CODEC = StreamCodec
         .composite(
-            SlotReferenceFactory.STREAM_CODEC, UseSlotReferencedItemPacket::slotReference,
+            PlayerSlotReference.STREAM_CODEC, UseSlotReferencedItemPacket::playerSlotReference,
             UseSlotReferencedItemPacket::new
         );
 
     public static void handle(final UseSlotReferencedItemPacket packet, final PacketContext ctx) {
         final Player player = ctx.getPlayer();
-        packet.slotReference.resolve(player).ifPresent(stack -> {
-            if (!(stack.getItem() instanceof SlotReferenceHandlerItem handlerItem)) {
-                return;
-            }
-            handlerItem.use((ServerPlayer) player, stack, packet.slotReference);
-        });
+        final ItemStack stack = packet.playerSlotReference.get(player);
+        if (!(stack.getItem() instanceof UsablePlayerSlotReferencedItem handlerItem)) {
+            return;
+        }
+        handlerItem.use((ServerPlayer) player, stack, packet.playerSlotReference);
     }
 
     @Override